extract some classes + add context clear
