+++ src/test/java/co/mvpmatch/backendtask1/web/api/ProductIntTest.kt
@@ -1,7 +1,7 @@
 package co.mvpmatch.backendtask1.web.api
 
 import co.mvpmatch.backendtask1.config.UserNotAllowedException
-import co.mvpmatch.backendtask1.helper.AuthenticationHelper
+import co.mvpmatch.backendtask1.helper.ContextHelper
 import co.mvpmatch.backendtask1.helper.ProductTestHelper
 import co.mvpmatch.backendtask1.helper.ProductTestHelper.Companion.testProductNameUpdated
 import org.junit.jupiter.api.Assertions
@@ -25,18 +25,18 @@ class ProductIntTest {
     protected lateinit var productController: ProductApiImpl
 
     @Autowired
-    private lateinit var authenticationHelper: AuthenticationHelper
+    private lateinit var contextHelper: ContextHelper
 
     @BeforeEach
     fun setup() {
         mockMvc = MockMvcBuilders.standaloneSetup(productController).build()
-        authenticationHelper.clearUsers()
+        contextHelper.clearData()
     }
 
     @Test
     fun `product crud is working`() {
         Assertions.assertNotNull(mockMvc)
-        val sellerToken = authenticationHelper.getMockSeller1Token()
+        val sellerToken = contextHelper.getMockSeller1Token()
         productHelper.createProduct(mockMvc!!, sellerToken)
         productHelper.testAndGetProduct(mockMvc!!, sellerToken)
         productHelper.updateProduct(mockMvc!!, sellerToken)
@@ -46,10 +46,10 @@ class ProductIntTest {
     @Test
     fun `only owner of product can modify it`() {
         Assertions.assertNotNull(mockMvc)
-        val seller1Token = authenticationHelper.getMockSeller1Token()
+        val seller1Token = contextHelper.getMockSeller1Token()
         productHelper.createProduct(mockMvc!!, seller1Token)
 
-        val seller2Token = authenticationHelper.getMockSeller2Token()
+        val seller2Token = contextHelper.getMockSeller2Token()
         assertThrows<UserNotAllowedException> {
             productHelper.updateProduct(mockMvc!!, seller2Token)
         }
@@ -58,7 +58,7 @@ class ProductIntTest {
     @Test
     fun `buyer is not allowed to add product`() {
         Assertions.assertNotNull(mockMvc)
-        val buyer1Token = authenticationHelper.getMockBuyer1Token()
+        val buyer1Token = contextHelper.getMockBuyer1Token()
         assertThrows<AccessDeniedException> {
             productHelper.createProduct(mockMvc!!, buyer1Token)
         }
